% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildTW_raster}
\alias{buildTW_raster}
\title{Build a Top Width Raster}
\usage{
buildTW_raster(r, nhd, TW)
}
\arguments{
\item{r}{Input raster with grid to output to}

\item{nhd}{A nhd sf polyline defining a river segment}

\item{TW}{a top width to buffer the nhd segment to}
}
\value{
a binary raster of in and out of bank areas
}
\description{
From a defined top width and NHD segement, create a raster defining in and out of bank regions. 
In bank = 1, out of bank = 0.
}
