% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_hand}
\alias{download_hand}
\title{Download HAND, Slope, and Cathmask rasters by HUC6}
\usage{
download_hand(HUC6, outdir, type = "all")
}
\arguments{
\item{HUC6}{The HUC6 unit(s) to download}

\item{outdir}{The folder path where data should be downloaded and extracted}

\item{type}{The type of data to download. Options include 'catchmask'. 'hand', or 'slp'. 
If 'all' is used (DEFAULT) all three type are downloaded.}
}
\description{
This function downloads staged HAND data to a local directory.
}
\examples{
\dontrun{
  download_hand("010100", outdir, "all")
}
}
