% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill.df}
\alias{fill.df}
\title{Fill Data.frame}
\usage{
fill.df(d, cm, pos = 1)
}
\arguments{
\item{d}{}

\item{cm}{a catchmask raster}

\item{pos}{in bank (1) or out of bank (0)}
}
\value{
data.frame
}
\description{
Fill Data.frame
}
\keyword{internal}
