% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_generation.R
\name{src_generate}
\alias{src_generate}
\title{Build Synthetic Rating Curve}
\usage{
src_generate(stages = NULL, catchmask, hand, slope, inbank, table,
  inN = 0.05, outN = 0.05, control = T)
}
\arguments{
\item{stages}{a vector of stages over which to predict a flow value}

\item{catchmask}{a catchment mask raster}

\item{hand}{a HAND raster}

\item{slope}{a slope raster}

\item{inbank}{an in-bank / out-bank raster}

\item{table}{a river-reach file}

\item{inN}{an in bank roughness value}

\item{outN}{an out of bank roughness value}

\item{control}{if TRUE in-bank N must be less then out of bank N}
}
\value{
a SRC data.frame
}
\description{
Build Synthetic Rating Curve
}
